requirements.txt
adbutils==0.6.4
aiohttp==3.6.2
apkutils2==1.0.0
appdirs==1.4.3
Appium-Python-Client==0.48
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
cigam==0.0.3
colorama==0.4.3
decorator==4.4.1
Deprecated==1.2.7
deprecation==2.0.7
Django==3.0
facebook-wda==0.4.1
humanize==0.5.1
idna==2.8
idna-ssl==1.1.0
image==1.5.27
logzero==1.5.0
lxml==4.4.2
multidict==4.5.2
numpy==1.18.0
packaging==19.2
Pillow==6.2.1
progress==1.5
py==1.8.0
pyee==6.0.0
pyelftools==0.26
PyMySQL==0.9.3
pyparsing==2.4.5
pyppeteer==0.0.25
pytz==2019.3
qrcode==6.1
requests==2.22.0
retry==0.9.2
retrying==1.3.3
selenium==3.141.0
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
tornado==6.0.3
tqdm==4.36.1
typing-extensions==3.7.4.1
uiautomator2==2.3.5.dev1
urllib3==1.25.6
websockets==6.0
weditor==0.4.2
whichcraft==0.6.1
wrapt==1.11.2
wxPython==4.0.7.post2
xmltodict==0.12.0
yarl==1.3.0
